+++ src/helpers/audioManager.js
@@ -12,7 +12,7 @@ import { withSessionRefresh } from "../lib/neonAuth";
 import { getBaseLanguageCode, validateLanguageForModel } from "../utils/languageSupport";
 import { classifyContext, getTargetAppInfo } from "../utils/contextClassifier";
 import { readCustomCleanupPrompt } from "../utils/promptStorage";
-import { buildCustomDictionaryPrompt } from "./customDictionaryPrompt";
+import { buildCustomDictionaryPrompt } from "./customDictionaryPrompt.mjs";
 import { normalizeAudioLevel } from "../utils/dictationWaveform.mjs";
 import { getReasoningAvailabilityCacheKey } from "../utils/reasoningAvailabilityCacheKey.mjs";
 import { resolveAsrFeatureFlags } from "../utils/asrFeatureFlags.mjs";